video: tegra: mode: Change tegra_dc_set_mode to only dirty if different
If the new mode is same as old mode, don't mark the mode as
dirty.  This is to handle the case during boot where bootloader
has already set a mode so we don't want the kernel to reset it.

If the mode is dirty, tegra_dc_program_mode() will be called,
which changes clocks and can cause a HPD drop and a visual glitch.

Change-Id: I380d41dff17eebc61ca20f2c0c7be97bb398fb69
Signed-off-by: Mike J. Chen <mjchen@google.com>